refactor CLI and renderer flow, add export safety coverage
