+++ test_cli.sh
@@ -27,6 +27,22 @@ assert_not_contains() {
     fi
 }
 
+assert_file_equals() {
+    file=$1
+    expected=$2
+    actual=$(cat "$file")
+    if [ "$actual" != "$expected" ]; then
+        fail "unexpected content in $file"
+    fi
+}
+
+assert_missing() {
+    path=$1
+    if [ -e "$path" ]; then
+        fail "did not expect $path to exist"
+    fi
+}
+
 TMPDIR=$(mktemp -d "${TMPDIR:-/tmp}/fuori-cli-test.XXXXXX")
 trap 'rm -rf "$TMPDIR"' EXIT INT TERM
 
@@ -41,6 +57,49 @@ assert_contains "$OUTSIDE/stdout.txt" "This document contains all the source cod
 assert_not_contains "$OUTSIDE/stderr.txt" "Git file-selection modes require"
 assert_not_contains "$OUTSIDE/stderr.txt" "git rev-parse failed"
 
+cat >"$OUTSIDE/notes.md" <<'EOF_NOTES'
+notes
+EOF_NOTES
+(cd "$OUTSIDE" && "$BIN" -o export.md >command_stdout.txt 2>command_stderr.txt)
+assert_contains "$OUTSIDE/export.md" "## main\\.c"
+assert_contains "$OUTSIDE/export.md" "## notes\\.md"
+assert_not_contains "$OUTSIDE/export.md" "export.md"
+assert_not_contains "$OUTSIDE/export.md" ".fuori.tmp."
+
+(cd "$OUTSIDE" && "$BIN" -o - >redirected.md 2>redirect_stderr.txt)
+assert_contains "$OUTSIDE/redirected.md" "## main\\.c"
+assert_contains "$OUTSIDE/redirected.md" "## notes\\.md"
+assert_not_contains "$OUTSIDE/redirected.md" "redirected.md"
+
+cat >"$OUTSIDE/existing.txt" <<'EOF_EXISTING'
+keep me
+EOF_EXISTING
+if (cd "$OUTSIDE" && "$BIN" -o existing.txt --no-clobber >/dev/null 2>no_clobber_stderr.txt); then
+    fail "expected --no-clobber to fail"
+fi
+assert_contains "$OUTSIDE/no_clobber_stderr.txt" "fuori: output file already exists: existing.txt"
+assert_file_equals "$OUTSIDE/existing.txt" "keep me"
+
+TOKEN_DIR="$TMPDIR/tokens"
+mkdir -p "$TOKEN_DIR"
+cat >"$TOKEN_DIR/big.txt" <<'EOF_BIG'
+This file is intentionally long enough to exceed a tiny token budget.
+EOF_BIG
+if (cd "$TOKEN_DIR" && "$BIN" --max-tokens 1 -o blocked.md >/dev/null 2>max_tokens_new_stderr.txt); then
+    fail "expected --max-tokens to fail for missing destination"
+fi
+assert_contains "$TOKEN_DIR/max_tokens_new_stderr.txt" "Error: estimated output is"
+assert_missing "$TOKEN_DIR/blocked.md"
+
+cat >"$TOKEN_DIR/blocked.md" <<'EOF_BLOCKED'
+original content
+EOF_BLOCKED
+if (cd "$TOKEN_DIR" && "$BIN" --max-tokens 1 -o blocked.md >/dev/null 2>max_tokens_existing_stderr.txt); then
+    fail "expected --max-tokens to fail for existing destination"
+fi
+assert_contains "$TOKEN_DIR/max_tokens_existing_stderr.txt" "Error: estimated output is"
+assert_file_equals "$TOKEN_DIR/blocked.md" "original content"
+
 REPO="$TMPDIR/repo"
 mkdir -p "$REPO/sub"
 (cd "$REPO" && git init -q)
@@ -50,7 +109,11 @@ EOF_IGNORE
 cat >"$REPO/sub/tracked.c" <<'EOF_TRACKED'
 int tracked(void) { return 1; }
 EOF_TRACKED
+cat >"$REPO/root_only.c" <<'EOF_ROOT'
+int root_only(void) { return 2; }
+EOF_ROOT
 (cd "$REPO" && git add .gitignore sub/tracked.c && \
+    git add root_only.c && \
     git -c user.name='fuori tests' -c user.email='fuori@example.com' commit -qm init)
 cat >"$REPO/sub/untracked.py" <<'EOF_UNTRACKED'
 print("hello")
@@ -65,6 +128,7 @@ assert_contains "$REPO/sub/stdout.txt" "├── tracked.c"
 assert_contains "$REPO/sub/stdout.txt" "└── untracked.py"
 assert_not_contains "$REPO/sub/stdout.txt" "ignored.log"
 assert_not_contains "$REPO/sub/stdout.txt" ".gitignore"
+assert_not_contains "$REPO/sub/stdout.txt" "root_only.c"
 
 (cd "$REPO" && "$BIN" --no-git -o - >stdout_no_git.txt 2>stderr_no_git.txt)
 assert_contains "$REPO/stdout_no_git.txt" "This document contains all the source code files from the current directory subtree using the local filesystem walker."
